+++ src/core/utils.ts
@@ -124,7 +124,7 @@ export async function queryInBatches<T>(
   take?: number,
 ): Promise<QueryResponse<T>> {
   if (take === undefined || take <= queryConfig.maxTakePerRequest) {
-    return await queryRecord(take || queryConfig.maxTakePerRequest);
+    return await queryRecord(take ?? queryConfig.maxTakePerRequest);
   }
 
   let queryResponse: T[] = [];
@@ -173,6 +173,107 @@ export async function queryInBatches<T>(
   return {
     data: queryResponse,
     totalCount,
+  };
+}
+
+/**
+ * Executes a query function repeatedly until continuation token is null i.e. all data is retrieved, adhering to the specified
+ * batch query configuration for maximum requests per second and items per request.
+ *
+ * @template T - The type of the data being queried.
+ * @param queryRecord - A function that performs the query. It takes the maximum number of items
+ *                      to retrieve (`take`) and an optional continuation token, and returns a
+ *                      promise that resolves to a `QueryResponse<T>`.
+ * @param config - The batch query configuration, including:
+ *   - `maxTakePerRequest`: The maximum number of items to retrieve per request.
+ *   - `requestsPerSecond`: The maximum number of requests to make per second.
+ * @returns A promise that containing all retrieved data
+ */
+export async function queryUntilComplete<T>(
+  queryRecord: (take: number, continuationToken?: string) => Promise<QueryResponse<T>>,
+  { maxTakePerRequest, requestsPerSecond }: BatchQueryConfig
+): Promise<QueryResponse<T>> {
+  const data: T[] = [];
+  let continuationToken: string | undefined;
+
+  do {
+    const start = Date.now();
+    for (let i = 0; i < requestsPerSecond; i++) {
+      try {
+        const response: QueryResponse<T> = await queryRecord(maxTakePerRequest, continuationToken);
+        data.push(...response.data);
+        continuationToken = response.continuationToken;
+
+        if (!continuationToken) {
+          break;
+        }
+      } catch (error) {
+        throw error; // Re-throw the error to be handled by the caller
+      }
+    }
+    const elapsed = Date.now() - start;
+    if (continuationToken && elapsed < 1000) {
+      await delay(1000 - elapsed);
+    }
+  } while (continuationToken);
+
+  return { data };
+}
+
+/**
+ * Executes a query function repeatedly until the response length < maxTakePerRequest i.e. all data is retrieved, adhering to the specified
+ * batch query configuration for maximum requests per second and items per request.
+ * 
+ * Note: This method should strictly be used with SLE APIs that require or support a skip parameter.
+ * 
+ * @param queryRecord - A function that performs the query. It takes the maximum number of items
+ *                      to retrieve (`take`) and an optional continuation token, and returns a
+ *                      promise that resolves to a `QueryResponse<T>`.
+ * @param config - The batch query configuration, including:
+ *   - `maxTakePerRequest`: The maximum number of items to retrieve per request.
+ *   - `requestsPerSecond`: The maximum number of requests to make per second.
+ * @returns A promise that containing all retrieved data
+ */
+export async function queryUsingSkip<T>(
+  queryRecord: (take: number, skip: number) => Promise<QueryResponse<T>>,
+  { maxTakePerRequest, requestsPerSecond }: BatchQueryConfig
+): Promise<QueryResponse<T>> {
+  const delay = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
+
+  let skip = 0;
+  const data: T[] = [];
+  let hasMore = true;
+
+  do {
+    const start = Date.now();
+
+    for (let i = 0; i < requestsPerSecond && hasMore; i++) {
+      try {
+        const response = await queryRecord(maxTakePerRequest, skip);
+        data.push(...response.data);
+
+        if (response.data.length < maxTakePerRequest) {
+          hasMore = false;
+          break;
+        }
+
+        skip += maxTakePerRequest;
+      } catch (error) {
+        console.error(`Error during batch fetch at skip=${skip}:`, error);
+        hasMore = false;
+        break;
+      }
+    }
+
+    const elapsed = Date.now() - start;
+    if (hasMore && elapsed < 1000) {
+      await delay(1000 - elapsed);
+    }
+  } while (hasMore);
+
+  return {
+    data,
+    totalCount: data.length,
   };
 }