+++ src/datasources/test-plans/components/query-builder/TestPlansQueryBuilder.test.tsx
@@ -2,14 +2,19 @@ import { QueryBuilderOption, Workspace } from 'core/types';
 import React, { ReactNode } from 'react';
 import { render } from '@testing-library/react';
 import { TestPlansQueryBuilder } from './TestPlansQueryBuilder';
+import { SystemAlias } from 'shared/types/QuerySystems.types';
 
 describe('TestPlansQueryBuilder', () => {
     let reactNode: ReactNode;
     const containerClass = 'smart-filter-group-condition-container';
     const workspace = { id: '1', name: 'Selected workspace' } as Workspace;
+    const systemAlias: SystemAlias = {
+        id: '1',
+        alias: 'System 1'
+    };
 
-    function renderElement(filter: string, workspaces: Workspace[] | null, globalVariableOptions: QueryBuilderOption[] = []) {
-        reactNode = React.createElement(TestPlansQueryBuilder, { filter, workspaces, globalVariableOptions, onChange: jest.fn() });
+    function renderElement(filter: string, workspaces: Workspace[] | null, systemAliases: SystemAlias[] | null, globalVariableOptions: QueryBuilderOption[] = []) {
+        reactNode = React.createElement(TestPlansQueryBuilder, { filter, workspaces, systemAliases, globalVariableOptions, onChange: jest.fn() });
         const renderResult = render(reactNode);
         return {
             renderResult,
@@ -18,16 +23,55 @@ describe('TestPlansQueryBuilder', () => {
     }
 
     it('should render empty query builder', () => {
-        const { renderResult, conditionsContainer } = renderElement('', []);
+        const { renderResult, conditionsContainer } = renderElement('', [], []);
 
         expect(conditionsContainer.length).toBe(1);
         expect(renderResult.findByLabelText('Empty condition row')).toBeTruthy();
     });
 
     it('should select workspace in query builder', () => {
-        const { conditionsContainer } = renderElement('workspace = "1"', [workspace]);
+        const { conditionsContainer } = renderElement('workspace = "1"', [workspace], []);
 
         expect(conditionsContainer?.length).toBe(1);
         expect(conditionsContainer.item(0)?.textContent).toContain(workspace.name);
     });
+
+    it('should select system alias in query builder', () => {
+        const { conditionsContainer } = renderElement('systemAliasName = "1"', [], [systemAlias]);
+
+        expect(conditionsContainer?.length).toBe(1);
+        expect(conditionsContainer.item(0)?.textContent).toContain(systemAlias.alias);
+    });
+
+
+    
+  [['${__from:date}', 'From'], ['${__to:date}', 'To'], ['${__now:date}', 'Now']].forEach(([value, label]) => {
+    it(`should select user friendly value for updated date`, () => {
+      const { conditionsContainer } = renderElement(`updatedAt > \"${value}\"`, [], []);
+
+      expect(conditionsContainer?.length).toBe(1);
+      expect(conditionsContainer.item(0)?.textContent).toContain(label);
+    });
+
+    it(`should select user friendly value for created date`, () => {
+        const { conditionsContainer } = renderElement(`created > \"${value}\"`, [], []);
+  
+        expect(conditionsContainer?.length).toBe(1);
+        expect(conditionsContainer.item(0)?.textContent).toContain(label);
+    });
+
+    it(`should select user friendly value for estimated end date`, () => {
+        const { conditionsContainer } = renderElement(`estimatedEndDate > \"${value}\"`, [], []);
+  
+        expect(conditionsContainer?.length).toBe(1);
+        expect(conditionsContainer.item(0)?.textContent).toContain(label);
+    });
+
+    it(`should select user friendly value for planned start date date`, () => {
+        const { conditionsContainer } = renderElement(`plannedStartDate > \"${value}\"`, [], []);
+  
+        expect(conditionsContainer?.length).toBe(1);
+        expect(conditionsContainer.item(0)?.textContent).toContain(label);
+    });
+  });
 });