+++ src/datasources/test-plans/components/query-builder/TestPlansQueryBuilder.tsx
@@ -2,13 +2,16 @@ import { SlQueryBuilder } from "core/components/SlQueryBuilder/SlQueryBuilder";
 import { queryBuilderMessages, QueryBuilderOperations } from "core/query-builder.constants";
 import { expressionBuilderCallback, expressionReaderCallback } from "core/query-builder.utils";
 import { QBField, QueryBuilderOption, Workspace } from "core/types";
+import { filterXSSField } from "core/utils";
 import { TestPlansQueryBuilderFields, TestPlansQueryBuilderStaticFields } from "datasources/test-plans/constants/TestPlansQueryBuilder.constants";
 import React, { useState, useEffect, useMemo } from "react";
+import { SystemAlias } from "shared/types/QuerySystems.types";
 import { QueryBuilderCustomOperation, QueryBuilderProps } from "smart-webcomponents-react/querybuilder";
 
 type TestPlansQueryBuilderProps = QueryBuilderProps & React.HTMLAttributes<Element> & {
     filter?: string;
     workspaces: Workspace[] | null;
+    systemAliases: SystemAlias[] | null;
     globalVariableOptions: QueryBuilderOption[];
 };
 
@@ -16,10 +19,24 @@ export const TestPlansQueryBuilder: React.FC<TestPlansQueryBuilderProps> = ({
     filter,
     onChange,
     workspaces,
+    systemAliases = [],
     globalVariableOptions,
 }) => {
     const [fields, setFields] = useState<QBField[]>([]);
-    const [operations, setOperations] = useState<QueryBuilderCustomOperation[]>([]);
+    const [operations, setOperations] = useState<QueryBuilderCustomOperation[]>([]); 
+    
+    const addOptionsToLookup = (field: QBField, options: QueryBuilderOption[]) => {
+        return {
+        ...field,
+        lookup: {
+            ...field.lookup,
+            dataSource: [
+            ...(field.lookup?.dataSource || []),
+            ...options,
+            ],
+        },
+        };
+    };
 
     const workspaceField = useMemo(() => {
         const workspaceField = TestPlansQueryBuilderFields.WORKSPACE;
@@ -27,24 +44,63 @@ export const TestPlansQueryBuilder: React.FC<TestPlansQueryBuilderProps> = ({
             return null;
         }
 
+        const options = workspaces.map(({ id, name }) => ({ label: name, value: id }))
+        return addOptionsToLookup(workspaceField, options);
+    }, [workspaces])
+
+    const systemAliasField = useMemo(() => {
+        const systemAliasField = TestPlansQueryBuilderFields.SYSTEM_ALIAS_NAME;
+        if (!systemAliases) {
+            return null;
+        }
+
         return {
-            ...workspaceField,
+            ...systemAliasField,
             lookup: {
-                ...workspaceField.lookup,
+                ...systemAliasField.lookup,
                 dataSource: [
-                    ...(workspaceField.lookup?.dataSource || []),
-                    ...workspaces.map(({ id, name }) => ({ label: name, value: id })),
+                    ...(systemAliasField.lookup?.dataSource || []),
+                    ...systemAliases.map(({ id, alias }) => ({ label: alias ?? '', value: id })),
                 ],
             },
         };
-    }, [workspaces])
+    }, [systemAliases])
+
+    const timeFields = useMemo(() => {
+        const timeOptions = [
+        { label: 'From', value: '${__from:date}' },
+        { label: 'To', value: '${__to:date}' },
+        { label: 'Now', value: '${__now:date}' },
+        ]
+
+        return [
+        addOptionsToLookup(TestPlansQueryBuilderFields.CREATED_AT, timeOptions),
+        addOptionsToLookup(TestPlansQueryBuilderFields.ESTIMATED_END_DATE, timeOptions),
+        addOptionsToLookup(TestPlansQueryBuilderFields.PLANNED_START_DATE, timeOptions),
+        addOptionsToLookup(TestPlansQueryBuilderFields.UPDATED_AT, timeOptions)
+        ]
+    }, []);
 
     useEffect(() => {
         if (!workspaceField) {
             return;
         }
 
-        const updatedFields = [...TestPlansQueryBuilderStaticFields, workspaceField]
+        if (!systemAliasField) {
+            return;
+        }
+
+        const updatedFields = [...TestPlansQueryBuilderStaticFields, ...timeFields, systemAliasField, workspaceField].map(field => {
+            if (field.lookup?.dataSource) {
+              return {
+                ...field,
+                lookup: {
+                  dataSource: [...globalVariableOptions, ...field.lookup?.dataSource].map(filterXSSField),
+                },
+              };
+            }
+            return field;
+        });
 
         setFields(updatedFields);
 
@@ -96,7 +152,7 @@ export const TestPlansQueryBuilder: React.FC<TestPlansQueryBuilderProps> = ({
 
         setOperations([...customOperations, ...keyValueOperations]);
 
-    }, [workspaceField]);
+    }, [workspaceField, systemAliasField, timeFields, globalVariableOptions]);
 
     return (
         <SlQueryBuilder