+++ src/datasources/test-plans/constants/TestPlansQueryBuilder.constants.ts
@@ -170,7 +170,10 @@ export const TestPlansQueryBuilderFields: Record<string, QBField> = {
             QueryBuilderOperations.DOES_NOT_EQUAL.name,
             QueryBuilderOperations.IS_BLANK.name,
             QueryBuilderOperations.IS_NOT_BLANK.name
-        ]
+        ],
+        lookup: {
+            dataSource: []
+        }
     },
     TEST_PLAN_ID: {
         label: 'Test plan ID',
@@ -233,23 +236,18 @@ export const TestPlansQueryBuilderFields: Record<string, QBField> = {
 
 export const TestPlansQueryBuilderStaticFields = [
     TestPlansQueryBuilderFields.ASSIGNED_TO,
-    TestPlansQueryBuilderFields.CREATED_AT,
     TestPlansQueryBuilderFields.CREATED_BY,
     TestPlansQueryBuilderFields.DESCRIPTION,
     TestPlansQueryBuilderFields.DUT_IDENTIFIER,
     TestPlansQueryBuilderFields.ESTIMATED_DURATION_IN_DAYS,
     TestPlansQueryBuilderFields.ESTIMATED_DURATION_IN_HOURS,
-    TestPlansQueryBuilderFields.ESTIMATED_END_DATE,
     TestPlansQueryBuilderFields.FIXTURE_IDENTIFIER,
     TestPlansQueryBuilderFields.NAME,
-    TestPlansQueryBuilderFields.PLANNED_START_DATE,
     TestPlansQueryBuilderFields.PRODUCT,
     TestPlansQueryBuilderFields.PROPERTIES,
     TestPlansQueryBuilderFields.STATE,
-    TestPlansQueryBuilderFields.SYSTEM_ALIAS_NAME,
     TestPlansQueryBuilderFields.TEST_PLAN_ID,
     TestPlansQueryBuilderFields.TEST_PROGRAM,
-    TestPlansQueryBuilderFields.UPDATED_AT,
     TestPlansQueryBuilderFields.UPDATED_BY,
     TestPlansQueryBuilderFields.WORK_ORDER_ID
 ];